// src/api/schemas/executionResponseDto.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Gourmet7 Agent API
 * AI Agent for DeFi operations
 * OpenAPI spec version: 1.0
 */
import type { ExecutionResponseDtoStatus } from './executionResponseDtoStatus';
import type { ExecutionSessionDto } from './executionSessionDto';

export interface ExecutionResponseDto {
  id: string;
  scheduleId: string;
  /** @nullable */
  sessionId?: string | null;
  session?: ExecutionSessionDto;
  status: ExecutionResponseDtoStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  error?: string | null;
  runNumber: number;
  createdAt: string;
}